backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pgvector
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
langgraph
langchain-ollama
langchain-core
langchain-community
langchain-groq
langchain-openai
langchain-google-genai
mcp
httpx
qdrant-client
python-dotenv
psutil
requests
GPUtil
ollama
rich
google-genai
aiosqlite
google-cloud-storage
